# docs/api/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Movscript API",
    "version": "1.0.0",
    "description": "Stable contract seed for critical Movscript backend APIs. Expand this file before wiring new generated frontend or plugin clients."
  },
  "servers": [
    {
      "url": "http://localhost:8765"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
...
      "Plugin": {
        "type": "object",
        "properties": {
          "ID": { "type": "integer" },
          "plugin_key": { "type": "string" },
          "name": { "type": "string" },
          "version": { "type": "string" },
          "enabled": { "type": "boolean" }
        },
        "required": ["ID", "plugin_key", "name", "version", "enabled"]
      },
      "PluginImport": {
        "type": "object",
        "properties": {
          "manifest_json": { "type": "string" },
          "path": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/plugins": {
      "get": {
        "operationId": "listPlugins",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Plugins", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Plugin" } } } } } }
      },
      "post": {
        "operationId": "importPlugin",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PluginImport" } } } },
        "responses": { "200": { "description": "Plugin", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Plugin" } } } }, "201": { "description": "Plugin", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Plugin" } } } } }
      }
    },
...
  }
}